Add files via upload
This script creates a managed instance in Azure for the current tenant subscriptions and for the resource group, vnet and subnet that is entered.  The user will name the MI, put in the admin and password, along with choose a region to create the parameters required for Azure.  No other steps are required.